myproject/requirements.txt
antiorm==1.2.1
asgiref==3.8.1
bcrypt==4.3.0
certifi==2025.1.31
cffi==1.17.1
channels==4.2.0
charset-normalizer==3.4.1
cryptography==44.0.2
dataclasses==0.6
db==0.1.1
db-sqlite3==0.0.1
Django==4.2
django-cors-headers==4.7.0
djangorestframework==3.14.0
djangorestframework_simplejwt==5.5.0
djongo==1.2.31
dnspython==2.7.0
ecdsa==0.19.1
idna==3.10
pillow==11.1.0
pyasn1==0.4.8
pycparser==2.22
PyJWT==2.9.0
pymongo==4.11.3
pyTelegramBotAPI==4.26.0
python-dotenv==1.1.0
python-jose==3.4.0
pytz==2025.2
requests==2.32.3
rsa==4.9
six==1.17.0
SQLite3-0611==0.0.1
sqlparse==0.5.3
telebot==0.0.5
tzdata==2025.1
urllib3==2.3.0
